Split ObjectOrErrorLogProcessor into separate ObjectLogProcessor and ErrorLogProcessor modules
