+++ lib/log-processors.js
@@ -9,7 +9,8 @@ const { MetadataLogProcessor } = require('./processors/MetadataLogProcessor')
 const { SemicolonLogProcessor } = require('./processors/SemicolonLogProcessor')
 const { TimeLogProcessor } = require('./processors/TimeLogProcessor')
 const { EOLLogProcessor } = require('./processors/EOLLogProcessor')
-const { ObjectOrErrorLogProcessor } = require('./processors/ObjectOrErrorLogProcessor')
+const { ErrorLogProcessor } = require('./processors/ErrorLogProcessor')
+const { ObjectLogProcessor } = require('./processors/ObjectLogProcessor')
 
 function createLogProcessor (definition) {
   if (definition instanceof LogProcessor) {
@@ -37,7 +38,8 @@ const builtInLogProcessors = {
   semicolon: SemicolonLogProcessor,
   message: MessageLogProcessor,
   eol: EOLLogProcessor,
-  objectOrError: ObjectOrErrorLogProcessor
+  error: ErrorLogProcessor,
+  object: ObjectLogProcessor
 }
 
 const defaultLogParsingSequence = [
@@ -50,7 +52,8 @@ const defaultLogParsingSequence = [
   builtInLogProcessors.semicolon,
   builtInLogProcessors.message,
   builtInLogProcessors.eol,
-  builtInLogProcessors.objectOrError
+  builtInLogProcessors.error,
+  builtInLogProcessors.object
 ]
 
 class State {
